queues/requirements.txt
python-dotenv
psycopg2
sqlalchemy
pika
sqlalchemy_utils